Remove `ShortCodeFormatter` file
All the methods in it were being used in only a single screen,
`ShortCodeSearchResultsScreen`. Moved the methods into that screen.
